beta to main fixxed

+++ src/main/java/net/emsee/thedungeon/entity/custom/interfaces/IMultiAttackAnimatedEntity.java
@@ -1,5 +1,5 @@
 package net.emsee.thedungeon.entity.custom.interfaces;
 
 public interface IMultiAttackAnimatedEntity {
-    void attackAnimation(byte id, byte version);
+    void attackAnimation(int id, byte version);
 }
